*refactor code on SmsHistory Page - separate router and page
